Moves scanner test alongside scanner caused by package change
